Fix item duct bug
This fixes a bug where ItemBuffer objects were shared by both the client
and server threads when playing singleplayer. This happened because the
messages for syncing buffers are not serialized in single player, so
they are never copied. This is usually not a bug because all of the
other messages use primitives or copy their data from their objects.